with double duration

+++ include/common_utilities/rviz_visualization.hpp
@@ -48,7 +48,7 @@ public:
      * @param duration in seconds
      */
     void publishMesh(Vector3d &pos, Vector4d& orientation, const char* modelname,
-                     const char *frame, const char *ns, int message_id, int duration);
+                     const char *frame, const char *ns, int message_id, double duration);
 
     /**
      * Publishes a sphere visualization marker
@@ -59,7 +59,7 @@ public:
      * @param rgda rgb color (0-1) plus transparancy
      * @param duration for this duration in seconds (0=forever)
      */
-    void publishSphere(Vector3d &pos, const char* frame, const char* ns, int message_id, COLOR color,float radius = 0.01, int duration=0);
+    void publishSphere(Vector3d &pos, const char* frame, const char* ns, int message_id, COLOR color,float radius = 0.01, double duration=0);
 
     /**
      * Publishes a cube visualization marker
@@ -71,7 +71,7 @@ public:
      * @param rgda rgb color (0-1) plus transparancy
      * @param duration for this duration in seconds (0=forever)
      */
-    void publishCube(Vector3d &pos, Vector4d &quat, const char* frame, const char* ns, int message_id, COLOR color,float radius = 0.01, int duration=0);
+    void publishCube(Vector3d &pos, Vector4d &quat, const char* frame, const char* ns, int message_id, COLOR color,float radius = 0.01, double duration=0);
 
     /**
      * Publishes a ray visualization marker
@@ -83,7 +83,7 @@ public:
      * @param color rgb color (0-1) plus transparancy
      * @param duration for this duration in seconds (0=forever)
      */
-    void publishRay(Vector3d &pos, Vector3d &dir, const char* frame, const char* ns, int message_id, COLOR color, int duration=0);
+    void publishRay(Vector3d &pos, Vector3d &dir, const char* frame, const char* ns, int message_id, COLOR color, double duration=0);
     /**
      * Publishes a text message marker
      * @param pos at this positon
@@ -95,7 +95,7 @@ public:
      * @param duration for this duration in seconds (0=forever)
      * @param height height of the text
      */
-    void publishText(Vector3d &pos, const char *text, const char *frame, const char *ns, int message_id, COLOR color, int duration, float height);
+    void publishText(Vector3d &pos, const char *text, const char *frame, const char *ns, int message_id, COLOR color, double duration, float height);
 private:
     ros::NodeHandlePtr nh;
     ros::Publisher visualization_pub;
